:seedling::droplet: ADDED class GithubLogin
... and moved previous spaghetti code into it.
